[SettingsMigration] Move the migration steps to their own .h/.cpp
